requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
sqlalchemy==2.0.35
alembic==1.13.3
psycopg2-binary==2.9.10
pydantic==2.9.2
pydantic-settings==2.6.1
jinja2==3.1.4
aiosmtpd==1.4.6
aiosmtplib==3.0.2
dkimpy==1.1.8
python-multipart==0.0.12
httpx==0.27.2
email-validator==2.2.0
cryptography==43.0.3
dnspython==2.7.0